Do not support subcommand "?" anymore
It worked as intended in tsvn.bat, but not in tsvn.sh if a file with a
filename with only one character existed in the current directory. The
"?" character was interpreted as file glob pattern, leading to the
following error (using filename "f"):

    Unknown subcommand: 'f'
    Type '/c/.../tsvn.sh' for usage.

But this does not even work with the "svn" command itself, even if
stated in the usage description.
Tested with Subversion 1.14.2 from TortoiseSVN package 1.14.3.

Without a matching filename ("?" evaluated as empty):

    Type 'svn help' for usage.

With a matching filename:

    Unknown subcommand: 'f'
    Type 'svn help' for usage.

After this change "tsvn.bat ?" always leads to this:

    Unknown subcommand: '?'
    Type 'C:\...\tsvn.bat' for usage.

Whereas "tsvn.sh ?" leads to one of these:

    Unknown subcommand: '?'
    Type '/C/.../tsvn.sh' for usage.

    Unknown subcommand: 'f'
    Type '/C/.../tsvn.sh' for usage.